% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBtoLinear.R
\name{dBtoLinear}
\alias{dBtoLinear}
\title{Convert deciBels scales to linear}
\usage{
dBtoLinear(x, factor="IL", ref=1)
}
\arguments{
\item{x}{Numeric. A numeric vector or a numeric matrix with dB values.}

\item{factor}{Character. Specify the factor to use for conversion. Use \code{SPL} (Sound Pressure Level) for amplitude-base data, with factor \code{20}, or \code{IL} (Intensity Level) for power-base data, with fator \code{10} (by default "IL").}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 1).}
}
\value{
The same input object with the values converted.
}
\description{
Function to convert decibels (dB) to linear values (µPa).
}
\details{
For further details on selecting the appropriate factor, we recommend consulting \href{https://dspillustrations.com/pages/posts/misc/decibel-conversion-factor-10-or-factor-20.html}{this} web page.
}
\examples{
dBtoLinear(c(80,60,65,62))
LineartodB(dBtoLinear(c(80,60,65,62)))

}
\seealso{
\code{\link{rms.dB}}, \code{\link{LineartodB}}, \code{\link[seewave]{convSPL}}
}
