% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqbands_cal.R
\name{leqbands_cal}
\alias{leqbands_cal}
\title{leqbands analysis for audiofiles with reference signal}
\usage{
leqbands_cal(
  files = "wd",
  channel = "left",
  from = 0,
  to = Inf,
  CalibPosition = NULL,
  CalibValue = NULL,
  ref = 20,
  weighting = "none",
  bands = "thirds",
  saveresults = F,
  outname = NULL,
  progressbar = T
)
}
\arguments{
\item{files}{Specifies audio file(s) to be analyzed. It can be set to "wd" to select all ".wav" files in the work directory, a single file name, a character vector with multiple file names, a Wave object, or a list of Wave objects (by default "wd"). Only ".wav" files are accepted.}

\item{channel}{Character. Choose “left” or “right” channel. Argument passed to \link[tuneR]{mono} function from \link[tuneR]{tuneR} (by default "left").}

\item{from}{Numeric. The start time (in seconds) of the segment to analyze. It can also be relative to the end of the file (in negative values). See examples for details.}

\item{to}{Numeric. The end time (in seconds) of the segment to analyze. It can also be relative to the beginning of the file (in negative values). See examples for details.}

\item{CalibPosition}{Numeric. Specifies the calibration position. It can be a negative (relative to the sound file duration) or a positive value, or a data.frame containing these combinations. This parameter is used in conjunction with \code{CalibValue}.}

\item{CalibValue}{Numeric. Specifies the calibration value. If \code{CalibPosition} is provided, it serves as the reference value. If \code{CalibPosition} is absent, \code{CalibValue} is used as the calibration reference.}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 20).}

\item{weighting}{Character. Defines the weighting curve for the analysis, passed to the \link[seewave]{dBweight} function. Accepted values are "A", "B", "C", "D", "ITU", and "none" (by default "none"). See \link[seewave]{dBweight} for details.}

\item{bands}{Character. Use "octaves" for octave bands or "thirds" for one-third octave bands intervals (by default "thirds").}

\item{saveresults}{Logical. Set to \code{TRUE} to save the results to a .txt file (by default \code{FALSE}).}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, specifies a name to append to the file name in the txt file (by default \code{NULL}).}

\item{progressbar}{Logical. Set \code{TRUE} to display a progress bar showing elapsed time and the last completed file number, or \code{FALSE} to hide it (default \code{TRUE}).}
}
\description{
This function passes the parameters to \code{\link{leqbands}} to automate the calibration process and return spectral analysis results in dB SPL.
}
\seealso{
\code{\link{leqbands}}
}
