% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LineartodB.R
\name{LineartodB}
\alias{LineartodB}
\title{Convert linear scales to deciBels}
\usage{
LineartodB(x, factor="IL", ref=1)
}
\arguments{
\item{x}{Numeric. A vector or matrix containing linear values (in µPa).}

\item{factor}{Character. Specify the factor for conversion. Use \code{SPL} (Sound Pressure Level) for amplitude-base data, with factor \code{20}, or \code{IL} (Intensity Level) for power-base data, with factor \code{10} (by default \code{IL}.)}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 1).}
}
\value{
The same input object with the values converted.
}
\description{
Function to convert linear values (µPa) to decibels (dB).
}
\details{
For details about the factor choice, we recommend consulting \href{https://dspillustrations.com/pages/posts/misc/decibel-conversion-factor-10-or-factor-20.html}{this} web page.
}
\examples{
dBtoLinear(c(80,60,65,62))
LineartodB(dBtoLinear(c(80,60,65,62)))

}
\seealso{
\code{\link{rms.dB}}, \code{\link{dBtoLinear}}, \code{\link[seewave]{convSPL}}
}
