% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propag.ruido.R
\name{propag.ruido}
\alias{propag.ruido}
\title{Sound decay models for points on GIS}
\usage{
propag.ruido(points=NA, ID=NULL, NIS0=NA, Alfa=NA, Beta=NA, elev.r=NA, contour.veg=NA, save.meta=T, name.meta=NULL, multicore=F, cl=detectCores()-1)
}
\arguments{
\item{points}{Matrix with longitude (first column) and latitude (second column) of each point to be modeled.}

\item{ID}{Matrix with IDs for each point the user wants to model. Do not use repeated names.}

\item{NIS0}{Sound intensity level at 1m. Parameter of the sound decay model. See details for explanations of the equation.}

\item{Alfa}{Matrix with values of  \emph{alfa} constant of the decay equation (see details for more information). Need to be one row for each point, even if they are the same.}

\item{Beta}{Same as the \emph{alfa} parameter, but representing values of \emph{beta} constant.}

\item{elev.r}{Raster with elevation values. OBS.: This raster will be used to extract all relevant GIS information (\emph{i. e.} reslution, size, crs).}

\item{contour.veg}{Object type 'SpatialPolygonDataFrame' with the contour of the open areas (areas without vegetagion cover). Still in improvement, need to work with multiple open areas.}

\item{save.meta}{Logical. If \code{TRUE} all steps of the process of modeling will be saved in your work directory. Objects will be saved with 'Results.propag_' at the beggining of the file name. (By default = \code{TRUE})}

\item{name.meta}{Optional. Define a custom name to be added after de default name if \code{save.meta} is \code{TRUE}.}

\item{multicore}{Logical. If \code{TRUE} multicore processing will be active. This parameter improve the processing time. (by deafult = \code{TRUE})}

\item{cl}{Numerical. Cluster object. Number of processes to be used in the \code{propag.ruido} computation. Only considered if \code{multicore} is \code{TRUE}. See documentation of package \code{\link{parallel}} and \code{\link{snow}} for details. (By default = number of CPU cores of the current host less one)}
}
\value{

}
\description{

}
\details{
Include here the equations used and their explanation.
}
\examples{

}
\references{

}
\seealso{
\code{\link{propag.ruido.linha}}
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>

Carlos Barros de Araújo <cabarau@gmail.com>
}
