% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rel.func.apply_internal.R
\name{rel.func.apply}
\alias{rel.func.apply}
\title{Internal noiseR function}
\usage{
rel.func.apply(x, pi.x, li.x = NULL, r, elev.rel, S.pi.x)
}
\description{
Internal \code{noiseR} function.

Function to improve the time computation of \code{\link{propag.ruido}}. It's only to be used inside an \code{\link[base]{apply}} family function.
}
\note{
These function is not to be called by the user.
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
