% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundmeter.R
\name{soundmeter}
\alias{soundmeter}
\title{Function that makes sound meter alike measurements}
\usage{
soundmeter(
  files = "wd",
  channel = "left",
  from = 0,
  to = Inf,
  CalibPosition = NULL,
  CalibValue = NULL,
  ref = 20,
  fw = "none",
  bands = "octaves",
  tw = "fast",
  saveresults = F,
  outname = NULL,
  progressbar = T
)
}
\arguments{
\item{files}{The audiofile to be analyzed. Can be "wd" to get all ".wav" files on the work directory, a file name (or a character containing a list of filenames) that exist in the work directory (only ".wav" files accepted), or an Wave object (or a list containing more than one Wave object). (By default: "wd")}

\item{channel}{Only "left" or "right" acepted. By default "left"}

\item{from}{Numeric. The start time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{to}{Numeric. The end time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{CalibPosition}{anda de mãos dadas com calib value. Pode ser negativo, positivo ou data.frame com essas combinações}

\item{CalibValue}{Anda de mãos dadas com calib position. Quando tem o position, ele é considerado o valor de referência, quando não tem o position, ele é considerado o valor de calibração.}

\item{ref}{Numerical. The reference value for dB conversion. For sound in water, the common is 1 microPa, and for sound on air 20 microPa. (By default 20)}

\item{fw}{Character. Argument passed to \code{\link[seewave]{dBweight}} to indicate the frequency weighting curve to use on the anlysis. 'A', 'B', 'C', 'D', 'ITU', and 'none' are supported. See \code{\link[seewave]{dBweight}} for details. (By default: "none")}

\item{bands}{Character. Choose the type of frequency band of the output. "octaves" to octaves bands intervals or "thirds" to one-third octaves bands intervals. (by deafault: "thirds")}

\item{tw}{Time weighting}

\item{saveresults}{Logical. Set \code{TRUE} if you want to save a txt file with the results of the function execution. (By default: \code{FALSE})}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, you can specify a name to appear on the txt file name after the default name. (By default: \code{NULL})}

\item{progressbar}{Logical. Activate or deactivate a progress bar with elapsed time and the last concluded file number. (By default: \code{TRUE})}
}
\value{
ESCREVER
}
\description{
ESCREVER
}
\details{
If your reference signal is in a separate file, we recommend you to get the \code{CalibValue} with \link{leqbands} function. It's examples provide more details.
}
\examples{
data("tham")
soundmeter(tham, CalibValue = 130.24, tw="slow") #slow time window with calib value
soundmeter(tham, CalibValue = 130.24, tw="fast") #fast

soundmeter(tham, CalibValue = 130.24, tw="fast", fw="A") #fast with frequency weight

soundmeter(tham, CalibValue = NULL, tw="fast", ref=1) #fast time window in dBFS

}
