% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.rasterdB.R
\name{sum.rasterdB}
\alias{sum.rasterdB}
\title{Sum raster with dB values}
\usage{
sum.rasterdB(x, na.rm=FALSE)
}
\arguments{
\item{x}{A list of raster objects with values in decibels (Intensity Level) to sum.}

\item{na.rm}{Logical. \code{NA} action, if \code{TRUE}, ignore \code{NA} values during the sum.}
}
\value{
One raster with the sum of all other raster of the input.
}
\description{
Function to sum rasters with decibel values (Intensity Level).
}
\details{
Include details.
}
\references{
Include citations.
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
