% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumdB.R
\name{sumdB}
\alias{sumdB}
\title{Sum with dB values}
\usage{
sumdB(x, level = "IL", na.rm = FALSE)
}
\arguments{
\item{x}{Numerical. A vector or a matrix with decibels values (dB).}

\item{level}{Character. Specify the factor for conversion. Use \code{SPL} (Sound Pressure Level) for amplitude-base data, with factor 20, or \code{IL} (Intensity Level) for power-base data, with fator 10 (By default \code{IL}).}

\item{na.rm}{Logical. Argument passed to \code{\link[base]{sum}}. If \code{TRUE} removes NA (by default \code{FALSE}).}
}
\value{
A numeric value representing the sum of x.
}
\description{
Sum with dB values
}
\details{
Function to compute the sum of decibels values (dB).

This function converts the decibels data to linear values using the \code{\link{dBtoLinear}} function, computes the sum, and then converts the result back to decibels using the \code{\link{LineartodB}} function.
}
\examples{
sumdB(c(80,60,65,62))
sumdB(c(30,30), level="IL")
sumdB(c(30,30), level="SPL")

}
\seealso{
\code{\link[seewave]{moredB}}
}
