% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tham-data.R
\docType{data}
\name{tham}
\alias{tham}
\title{\emph{Thamnophilus stictocephalus} song in a landscape}
\format{
An object of class \code{"Wave"}; see \link[tuneR]{readWave}.
}
\source{
Recording by Ingrid Maria Denóbile Torres
}
\usage{
data(tham)
}
\description{
\emph{Thamnophilus stictocephalus} song in a landscape
}
\details{
A landscape recording presenting \emph{Thamnophilus stictocephalus} song between 3.883035 and 7.044417 seconds. The "calib.value" for this record is 130.24.
}
\keyword{datasets}
