% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timbreCal.R
\name{timbreCal}
\alias{timbreCal}
\title{Timbre analysis for audiofiles with reference signal}
\usage{
timbreCal(files="wd", SignalDur=NULL, RefValue=NULL, ref=20, weighting="none",
       bands="thirds", saveresults=F, outname=NULL, time.mess=T, stat.mess=T)
}
\arguments{
\item{files}{The audiofile to be analyzed. Can be "wd" to get all ".wav" files on the work directory, a file name (or a character containing a list of filenames) that exist in the work directory (only ".wav" files accepted), or an Wave object (or a list containing more than one Wave object). (By default: "wd")}

\item{SignalDur}{Numerical. Specify the reference signal duration (in seconds) on the beggining of the audiofile. (By default: \code{NULL})}

\item{RefValue}{Numerical. Specify the reference signal sound pressure level (in deciBells SPL) on the beggining of the audiofile. (By default: \code{NULL})}

\item{ref}{Numerical. The reference value for dB conversion. For sound in water, the common is 1 microPa, and for sound on air 20 microPa. (By default 20)}

\item{weighting}{Character. Indicate the weighting curve to use on the anlysis. A, B, C and none are supported. (By default: "none")}

\item{bands}{Character. Choose the type of frequency band of the output. "octaves" to octaves bands intervals or "thirds" to one-third octaves bands intervals. (by deafault: "thirds")}

\item{saveresults}{Logical. Set \code{TRUE} if you want to save a txt file with the results of the function execution. (By default: \code{FALSE})}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, you can specify a name to appear on the txt file name after the default name. (By defaulf: \code{NULL})}

\item{time.mess}{Logical. Activate or deactivate message of time to complete the function execution. (By default: \code{TRUE})}

\item{stat.mess}{Logical. Activate or deactivate status message of the function execution. (By default: \code{TRUE})}
}
\description{
This function passes the parameters to \code{\link{timbre}} to automatize the calibration and return spectral analysis with dB SPL results.
}
\details{
To use this function, the audio file must begin with 2 seconds of silence, followed by a reference signal with known SPL, followed by another 2 seconds of silence, and the following sound to analyze.

The duration of the reference signal must be specified (in seconds) on the \code{SignalDur} argument and his value (in dB SPL) on the \code{refValue} argument.
}
\seealso{
\code{\link{timbre}}
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
