% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calibration}
\alias{calibration}
\title{Extract calbration factor from reference sinal}
\usage{
calibration(
  file = NULL,
  channel = "left",
  from = 0,
  to = Inf,
  leqsignal = NA,
  weighting = "none",
  ref = 20
)
}
\arguments{
\item{file}{Specifies audio file to be analyzed. It can be a single file name on the computer or a Wave object on the environment (by default NULL). Only ".wav" files are accepted.}

\item{channel}{Character. Choose “left” or “right” channel (by default "left"). Argument passed to \link[tuneR]{mono} function from \link[tuneR]{tuneR} to extract the desired channel.}

\item{from}{Numeric. The start time (in seconds) of the segment to analyze. It can also be relative to the end of the file (in negative values). See examples for details.}

\item{to}{Numeric. The end time (in seconds) of the segment to analyze. Could also be relative to the beginning of the file (in negative values). See examples for details.}

\item{weighting}{Character. Defines the weighting curve for the analysis, passed to the \code{\link[seewave]{dBweight}} function. Accepted values are "A", "B", "C", "D", "ITU", and "none" (by default "none"). See \code{\link[seewave]{dBweight}} for details.
#' @param leqsignal Numeric. Specifies the sound pressure level (in dB SPL) for the signal in the audio (by default \code{NA}).}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 20).}
}
\description{
ESCREVER
}
