% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBtoLinear&LineartodB.R
\name{dBtoLinear}
\alias{dBtoLinear}
\title{Convert deciBels scales to linear}
\usage{
dBtoLinear(x, factor="IL", ref=1)
}
\arguments{
\item{x}{Numerical. A numeric vector or a numeric matrix with dB values for \code{dBtoLinear} funcion or linear values for \code{\link{LineartodB}}.}

\item{factor}{Character. Specify in what factor the function should use to convert your data. \code{SPL} (Sound Pressure Level) for amplitude like data (factor \code{20}) or \code{IL} (Intensity Level) for power like (fator \code{10}). (By default \code{IL})}

\item{ref}{Numerical. Reference value for conversion. For Sound in water the ref is 1 microPa and on air 20 microPa. (By default 1)}
}
\value{
The same object of the input with the converted values.
}
\description{
Function to convert dB scales. The conversion can be made either from dB to linear (\code{dBtoLinear}) or from linear to dB (\code{\link{LineartodB}}).
}
\details{
For details about the factor choice, we recommend the reading of \href{https://dspillustrations.com/pages/posts/misc/decibel-conversion-factor-10-or-factor-20.html}{this} web page.
}
\examples{
dBtoLinear(c(80,60,65,62))
LineartodB(dBtoLinear(c(80,60,65,62)))

}
\seealso{
\code{\link{rms.dB}}, \code{\link{LineartodB}}, \code{\link[seewave]{convSPL}}
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
