% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq.bands.R
\name{freq.bands}
\alias{freq.bands}
\title{Interval Limits}
\usage{
freq.bands(x, interval = 2, below = 3, above = 3)
}
\arguments{
\item{x}{Numerical. The frequency values to start the bands' calculation.}

\item{interval}{Numeric. The interval pattern to be applied. See details for more.}

\item{below}{Numerical. Number of intervals below x.}

\item{above}{Numerical. Number of intervals above x.}
}
\value{
A numeric vector with the frequency limits of each interval.
}
\description{
Function to compute intervals from patterns defined by the user. This function was adapted from \code{\link[seewave]{octaves}} functions from \code{\link[seewave]{seewave}} package.
}
\details{
The interval specified is applied by x/interval (for bellow) of x*interval (for upper). Some examples of intervals that can be applied are:
\itemize{
  \item Octaves = 2
  \item Third of octaves = 2^(1/3)
  \item Perfect fifth (music theory) = 3/2
  \item Major third (music theory) = 5/4
  \item Minor third (music theory) = 6/5
  \item This \link[https://academics.hamilton.edu/music/spellman/class_notes/music_theory.htm]{link} shows other values that can be used.
}
}
\examples{
freq.bands(1000, interval=2, below = 1, above = 1) #octaves
freq.bands(1000, interval=2^(1/3), below = 3, above = 3) #Third of octaves

#https://academics.hamilton.edu/music/spellman/class_notes/music_theory.htm
freq.bands(440, interval=3/2, below = 0, above = 1) #Perfect fifth (music theory) of A 440
freq.bands(440, interval=5/4, below = 0, above = 1) #Major third (music theory) of A 440
freq.bands(440, interval=6/5, below = 0, above = 1) #Minor third (music theory) of A 440

}
\seealso{
\code{\link[seewave]{octaves}}
}
