% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqbands.R
\name{leqbands}
\alias{leqbands}
\title{Equivalent Level per octaves or one-thirds octaves}
\usage{
leqbands(
  files = "wd",
  channel = "left",
  from = 0,
  to = Inf,
  weighting = "none",
  bands = "thirds",
  ref = 20,
  Leq.calib = NULL,
  Calib.value = NULL,
  saveresults = F,
  outname = NULL,
  progressbar = T
)
}
\arguments{
\item{files}{The audiofile to be analyzed. Can be "wd" to get all ".wav" files on the work directory, a file name (or a character containing a list of filenames) that exist in the work directory (only ".wav" files accepted), or an Wave object (or a list containing more than one Wave object). (By default: "wd")}

\item{channel}{Character. Choose “left” or “right” channel. Argument passe to \link[tuneR]{mono} function from \link[tuneR]{tuneR} to extract the desired channel.}

\item{from}{Numeric. The start time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{to}{Numeric. The end time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{weighting}{Character. Argument passed to \code{\link[seewave]{dBweight}} to indicate the weighting curve to use on the anlysis. 'A', 'B', 'C', 'D', 'ITU', and 'none' are supported. See \code{\link[seewave]{dBweight}} for details. (By default: "none")}

\item{bands}{Character. Choose the type of frequency band of the output. "octaves" to octaves bands intervals or "thirds" to one-third octaves bands intervals. (by deafault: "thirds")}

\item{ref}{Numerical. The reference value for dB conversion. For sound in water, the common is 1 microPa, and for sound on air 20 microPa. (By default 20)}

\item{Leq.calib}{Numerical. The sound pressure level (in dB SPL) that the signal in the audio file must have (by default: \code{NULL}). Can not be set if \code{Calib.value} is also set.}

\item{Calib.value}{Numerical. The calibration value returned from the analysis of a reference signal using \code{Leq.calib} (by default: \code{NULL}). Can not be set if \code{Leq.calib} is also set.}

\item{saveresults}{Logical. Set \code{TRUE} if you want to save a txt file with the results of the function execution. (By default: \code{FALSE})}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, you can specify a name to appear on the txt file name after the default name. (By default: \code{NULL})}

\item{progressbar}{Logical. Activate or deactivate a progress bar with elapsed time and the last concluded file number. (By default: \code{TRUE})}
}
\value{
ESCREVER
}
\description{
ESCREVER
}
\details{
Caution: You need to use an audiofile with entire values of seconds of duration to avoid bugs. Example: 35s, 60s, 19s. By default, the function will trunc your audiofile to the next entire value of seconds.

These function works only with mono audiofiles.

The audio files need to have at least 44100Hz of sampling rate.

If you intend to work with decibels at full scale (dBFS), we recommend setting \code{ref=1}. With this, your results will be relative to 0 dBFS.
}
\examples{
data(tham)
leqbands(tham)
leqbands(tham, Calib.value=130.24)
leqbands(tham, Calib.value=130.24, weighting="A")
leqbands(tham, Calib.value=130.24, weighting="A", bands="octaves")

leqbands(tham, Leq.calib=48.17, weighting="A")

leqbands(tham, from=-3.49, to=-0.9, ref=1) #dB at Full Scale
leqbands(tham, from=-3.49, to=-0.9, Calib.value=130.24) #song Sound Pressure Level
leqbands(tham, from=0, to=3.8, Calib.value=130.24) #background Sound Pressure Level



}
\references{
Power spectrum adapted from: Carcagno, S. 2013. Basic Sound Processing with R [Blog post]. Retrieved from http://samcarcagno.altervista.org/blog/basic-sound-processing-r/

Miyara, F. 2017. Software-Based Acoustical Measurements. Springer. 429 pp. DOI: 10.1007/978-3-319-55871-4
}
