% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leqbands.R
\name{leqbands}
\alias{leqbands}
\title{Equivalent Level per octaves or one-third octaves}
\usage{
leqbands(
  files = "wd",
  channel = "left",
  from = 0,
  to = Inf,
  weighting = "none",
  bands = "thirds",
  ref = 20,
  Leq.calib = NULL,
  Calib.value = NULL,
  saveresults = F,
  outname = NULL,
  progressbar = T
)
}
\arguments{
\item{files}{Specifies audio file(s) to be analyzed. It can be set to "wd" to select all ".wav" files in the work directory, a single file name, a character vector with multiple file names, a Wave object, or a list of Wave objects (by default "wd"). Only ".wav" files are accepted.}

\item{channel}{Character. Choose “left” or “right” channel (by default "left"). Argument passed to \link[tuneR]{mono} function from \link[tuneR]{tuneR} to extract the desired channel.}

\item{from}{Numeric. The start time (in seconds) of the segment to analyze. It can also be relative to the end of the file (in negative values). See examples for details.}

\item{to}{Numeric. The end time (in seconds) of the segment to analyze. Could also be relative to the beginning of the file (in negative values). See examples for details.}

\item{weighting}{Character. Defines the weighting curve for the analysis, passed to the \code{\link[seewave]{dBweight}} function. Accepted values are "A", "B", "C", "D", "ITU", and "none" (by default "none"). See \code{\link[seewave]{dBweight}} for details.}

\item{bands}{Character. Use "octaves" for octave bands or "thirds" for one-third octave bands intervals (by default "thirds").}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 20).}

\item{Leq.calib}{Numeric. Specifies the sound pressure level (in dB SPL) for the signal in the audio (by default \code{NULL}). Cannot be set if \code{Calib.value} is specified.}

\item{Calib.value}{Numeric. Specifies the calibration value (by default \code{NULL}). Can not be set if \code{Leq.calib} is specified.}

\item{saveresults}{Logical. Set to \code{TRUE}  to save the results to a .txt file (by default \code{FALSE}).}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, specifies a name to append to the file name in the txt file (by default \code{NULL}).}

\item{progressbar}{Logical. Set \code{TRUE} to display a progress bar showing elapsed time and the last completed file number, or \code{FALSE} to hide it (default \code{TRUE}).}
}
\description{
Function to comput the Equivalent Level (Leq) across octaves or one-third octaves.
}
\details{
Caution: Ensure that the audio file has a duration with whole second values to avoid bugs. For example: 35s, 60s, 19s. By default, the function will truncate your audio file to the next whole second.

This function only works with mono audio files, and the audio must have a sampling rate of at least 44,100 Hz.

If you are working with decibels at full scale (dBFS), we recommend setting \code{ref=1} so the results will be relative to 0 dBFS.
}
\examples{
data(tham)
leqbands(tham)
leqbands(tham, Calib.value=130.24)
leqbands(tham, Calib.value=130.24, weighting="A")
leqbands(tham, Calib.value=130.24, weighting="A", bands="octaves")

leqbands(tham, Leq.calib=48.17, weighting="A")

leqbands(tham, from=-3.49, to=-0.9, ref=1) #dB at Full Scale
leqbands(tham, from=-3.49, to=-0.9, Calib.value=130.24) #song Sound Pressure Level
leqbands(tham, from=0, to=3.8, Calib.value=130.24) #background Sound Pressure Level



}
\references{
Power spectrum adapted from: Carcagno, S. 2013. Basic Sound Processing with R [Blog post]. Retrieved from http://samcarcagno.altervista.org/blog/basic-sound-processing-r/

Miyara, F. 2017. Software-Based Acoustical Measurements. Springer. 429 pp. DOI: 10.1007/978-3-319-55871-4
}
