% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dBtoLinear&LineartodB.R
\name{LineartodB}
\alias{LineartodB}
\title{Convert linear scales to deciBels}
\usage{
LineartodB(x, factor="IL", ref=1)
}
\arguments{
\item{x}{Numerical. A numeric vector or a numeric matrix with dB values for \code{\link{dBtoLinear}} funcion or linear values for \code{LineartodB}.}

\item{factor}{Character. Specify in what factor the functoin should use to convert your data. \code{SPL} (Sound Pressute Level) for amplitude like data (factor \code{20}) or \code{IL} (Intensity Level) for power like (fator \code{10}). (By default \code{IL})}

\item{ref}{Numerical. Reference value for conversion. For Sound in water the ref is 1microPa and on air 20 microPa. (By default 1)}
}
\value{
The same object of the input with the converted values.
}
\description{
Function to convert dB scales. The conversion can be made either from dB to linear(\code{\link{dBtoLinear}}) or from linear to dB (\code{LineartodB}).
}
\details{
For details about the factor choice, we recommend the reading of \href{https://dspillustrations.com/pages/posts/misc/decibel-conversion-factor-10-or-factor-20.html}{this} web page.
}
\examples{
dBtoLinear(c(80,60,65,62))
LineartodB(dBtoLinear(c(80,60,65,62)))

}
\seealso{
\code{\link{rms.dB}}, \code{\link{dBtoLinear}}, \code{\link[seewave]{convSPL}}
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
