% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwrspec.R
\name{pwrspec}
\alias{pwrspec}
\title{Power Spectrum of a sound file}
\usage{
pwrspec(
  file,
  channel = "left",
  from = 0,
  to = Inf,
  bandpass = c(0, Inf),
  res.scale = "microPa",
  ref = 1
)
}
\arguments{
\item{file}{A wave file path or a class Wave object already loaded into R environment.}

\item{channel}{Character. Choose “left” or “right” channel. Argument passed to \link[tuneR]{mono} function from \link[tuneR]{tuneR} (by default "left").}

\item{from}{Numeric. The start time (in seconds) of the segment to analyze. It can also be relative to the end of the file (in negative values). See examples for details.}

\item{to}{Numeric. The end time (in seconds) of the segment to analyze. It can also be relative to the beginning of the file (in negative values). See examples for details.}

\item{bandpass}{A vector of length two specifying the lower and upper limits of the bandpass interval in Hz.}

\item{res.scale}{Character. Specifies the scale for adjusting the power spectrum amplitude. Use \code{microPa} for linear values in µPa, or \code{dB} for decibel values in dB SPL (by default \code{microPa}).}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 1).}
}
\value{
A data.frame with columns:
\itemize{
  \item Freq.Hz: numeric, frequency in Hertz
  \item Amp.microPa: numeric, amplitude in µPa (present when res.scale = "microPa")
  \item Amp.dB: numeric, amplitude in dB SPL (present when res.scale = "dB")
}
}
\description{
This functions computes a power spectrum from a sound file.
}
\examples{
data(tham)
pwrspec(tham)

#with from, to and bandpass
pwrspec(tham, from=3.8, to=7.1, bandpass=c(900,2000))

#from and to relative to duration (negatives values from the end of the soundfiles)
pwrspec(tham, from=-3.49, to=-0.9, bandpass=c(900,2000))


}
\references{
Power spectrum adapted from: Carcagno, S. 2013. Basic Sound Processing with R [Blog post]. Retrieved from http://samcarcagno.altervista.org/blog/basic-sound-processing-r/

Miyara, F. 2017. Software-Based Acoustical Measurements. Springer. 429 pp. DOI: 10.1007/978-3-319-55871-4
}
