% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwrspec.R
\name{pwrspec}
\alias{pwrspec}
\title{Power Spectrum of a sound file}
\usage{
pwrspec(file, from = 0, to = Inf, bandpass = c(0, Inf))
}
\arguments{
\item{file}{A wave file path in your computer or a class Wave object already loaded into R environment.}

\item{from}{Numeric. The start time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{to}{Numeric. The end time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{bandpass}{A vector with length two with lower and upper limits of the band pass interval in Hz.}
}
\value{
This function returns a data.frame with Frequency (Hz) and Intensity (microPa) of the file.
}
\description{
This functions computes a power spectrum from a sound file.
}
\examples{
data(tham)
pwrspec(tham)

#with from, to and bandpass
pwrspec(tham, from=3.8, to=7.1, bandpass=c(900,2000))

#from and to relative to duration (negatives values from the end of the soundfiles)
pwrspec(tham, from=-3.49, to=-0.9, bandpass=c(900,2000))


}
\references{
Power spectrum adapted from: Carcagno, S. 2013. Basic Sound Processing with R [Blog post]. Retrieved from http://samcarcagno.altervista.org/blog/basic-sound-processing-r/

Miyara, F. 2017. Software-Based Acoustical Measurements. Springer. 429 pp. DOI: 10.1007/978-3-319-55871-4
}
