% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.dB.R
\name{rms.dB}
\alias{rms.dB}
\title{Root Mean Square with dB values}
\usage{
rms.dB(x, level = "SPL", na.rm = FALSE)
}
\arguments{
\item{x}{Numeric. A vector or a matrix with decibels values (dB).}

\item{level}{Character. Specify the factor for conversion. Use \code{SPL} (Sound Pressure Level) for amplitude-base data, with factor 20, or \code{IL} (Intensity Level) for power-base data, with fator 10 (By default \code{SPL}).}

\item{na.rm}{Logical. Argument passed to \code{\link[base]{mean}}. If \code{TRUE} removes NA (By default \code{FALSE}).}
}
\value{
A numeric value that representing the root mean square of x.
}
\description{
Function to compute the root mean square (RMS) of values in decibels (dB).
}
\details{
This function converts the decibels data to linear values using the \code{\link{dBtoLinear}} function, computes the Root Mean Square (rms), and then converts the result back to decibels using the \code{\link{LineartodB}} function.

This function s adapted from the \code{\link[seewave]{meandB}} and \code{\link[seewave]{rms}} functions from \code{\link[seewave]{seewave}} package. See their documentation for more details.
}
\examples{
rms.dB(c(80,60,65,62))

}
\seealso{
\code{\link[seewave]{meandB}}, \code{\link[seewave]{rms}}
}
