% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.dB.R
\name{rms.dB}
\alias{rms.dB}
\title{Root Mean Square with dB values}
\usage{
rms.dB(x, level = "SPL", na.rm = FALSE)
}
\arguments{
\item{x}{Numerical. A numeric vector or a numeric matrix with dB values.}

\item{level}{Character. Specify in what scale your data is. \code{SPL} for Sound Pressute Level or \code{IL} for Intensity Level. (By default \code{SPL})}

\item{na.rm}{Logical. Argument passed to \code{\link[base]{mean}}. Should NA be removed? (By default FALSE)}
}
\value{
A numeric value that represents the root mean square of x.
}
\description{
Function to compute the root mean square (RMS) of values in decibels (dB).
}
\details{
This function converts your dB data to linear values (through \code{\link{dBtoLinear}} function), compute the Root Mean Square (rms), and converts the result back to dB (through \code{\link{LineartodB}} function).

This function was adapted from \code{\link[seewave]{meandB}} and \code{\link[seewave]{rms}} functions from \code{\link[seewave]{seewave}} package. See their help for more details.
}
\examples{
rms.dB(c(80,60,65,62))

}
\seealso{
\code{\link[seewave]{meandB}}, \code{\link[seewave]{rms}}
}
