% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rms.dB.R
\name{rms.dB}
\alias{rms.dB}
\title{Root Mean Square with dB values}
\usage{
rms.dB(x, level="SPL", ...)
}
\arguments{
\item{x}{Numerical. A numeric vector or a numeric matrix with dB values.}

\item{level}{Character. Specify in what scale your data is. \code{SPL} for Sound Pressute Level or \code{IL} for Intensity Level.}

\item{...}{Further arguments passed to \code{\link[base]{mean}}.}
}
\value{
A numeric value that represent the root mean square of x.
}
\description{
Function to compute the root mean square (RMS) of values in decibels (dB).
}
\details{
This function convert your dB data to linear values, compute the Root Mean Square (rms) and converts the result to dB again.

This function was adapted from \code{\link[seewave]{meandB}} and \code{\link[seewave]{rms}} functions from \code{\link[seewave]{seewave}} package. See their help for more details.
}
\seealso{
\code{\link[seewave]{meandB}}, \code{\link[seewave]{rms}}
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
