% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song.level.R
\name{song.level}
\alias{song.level}
\title{RMS from a sample of a sound file}
\usage{
song.level(
  files = "wd",
  channel = "left",
  from = 0,
  to = Inf,
  freq.interval = c(0, Inf),
  fdom.int = c(0, Inf),
  wl = 512,
  ovlp = 50,
  CalibPosition = NULL,
  CalibValue = NULL,
  freq.weight = "none",
  ref = 20,
  progressbar = T
)
}
\arguments{
\item{files}{Specifies audio file(s) to be analyzed. It can be set to "wd" to select all ".wav" files in the work directory, a single file name, a character vector with multiple file names, a Wave object, or a list of Wave objects (by default "wd"). Only ".wav" files are accepted.}

\item{channel}{Character. Choose “left” or “right” channel. Argument passed to \link[tuneR]{mono} function from \link[tuneR]{tuneR} (by default "left").}

\item{from}{Numeric. Specifies the start time (in seconds) of the segment to analyze. It can also be relative to the end of the file (in negative values). See examples for details.}

\item{to}{Numeric. Specifies the end time (in seconds) of the segment to analyze. It can also be relative to the beginning of the file (in negative values). See examples for more details.}

\item{freq.interval}{Specifies the frequency interval for computing the RMS. It can be a vector of length two, with the lower and upper frequency bounds (in Hz), or a pattern to calculate a interval (e.g., octaves). For more details, refer to \link{freq.bands}.}

\item{fdom.int}{A vector of length two specifying the lower and upper frequency bounds (in Hz) to determine the dominant frequency. This frequency will be used as the center of the interval only if a pattern is specified in \code{freq.interval}.}

\item{wl}{Numeric. Window length for the analysis. It must be a even number of points (by default 512).}

\item{ovlp}{Numeric. Overlap percentage between two successive windows (by default 50). Argument passed to \link[seewave]{meanspec} function from \link[seewave]{seewave}.}

\item{CalibPosition}{Numeric. Specifies the calibration position. It can be a negative (relative to the sound file duration) or a positive value, or a data.frame containing these combinations. This parameter is used in conjunction with \code{CalibValue}.}

\item{CalibValue}{Numeric. Canbe a value to apply of the ref value from a calib signal (specified by CalibPosition).}

\item{freq.weight}{Character. Defines the weighting curve for the analysis, passed to the \link[seewave]{dBweight} function. Accepted values are "A", "B", "C", "D", "ITU", and "none" (by default "none"). See \link[seewave]{dBweight} for details.}

\item{ref}{Numeric. The reference value for dB conversion. For sound in water, the common reference is 1 µPa, and for sound in air,it is 20 µPa (by default 20).}

\item{progressbar}{Logical. Set to \code{TRUE} to display a progress bar showing elapsed time and the last completed file number, or \code{FALSE} to hide it (default \code{TRUE}).}
}
\description{
RMS (Root Mean Square) from a sample of a sound file.
}
\examples{
song.level(tham, freq.interval=c(22, 20000))

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24)

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24, freq.weight="A")

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24, freq.weight="B")

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24, freq.weight="C")

song.level(tham, from = 3.883035, to=7.044417, freq.interval=c(22.09429, 22627.38), CalibValue = 130.24, freq.weight="A")

#Perfect fifth (music theory)
song.level(tham, fdom.int = c(800, 2000), from = 3.8, to=7, freq.interval=3/2, CalibValue = 130.24, freq.weight="A")

#Perfect fifth (music theory)
song.level(tham, fdom.int = c(800, 2000), from = 3.8, to=7, freq.interval=3/2, CalibValue = 130.24, freq.weight="A")

}
