% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/song.level.R
\name{song.level}
\alias{song.level}
\title{RMS from a sample of a sound file}
\usage{
song.level(
  files = "wd",
  channel = "left",
  from = 0,
  to = Inf,
  freq.interval = c(0, Inf),
  fdom.int = c(0, Inf),
  wl = 512,
  ovlp = 50,
  CalibPosition = NULL,
  CalibValue = NULL,
  freq.weight = "none",
  ref = 20,
  progressbar = T
)
}
\arguments{
\item{files}{The audiofile to be analyzed. Can be "wd" to get all ".wav" files on the work directory, a file name (or a character containing a list of filenames) that exist in the work directory (only ".wav" files accepted), or an Wave object (or a list containing more than one Wave object). (By default: "wd")}

\item{channel}{Argument passed to \link[tuneR]{mono} function from \link[tuneR]{tuneR} to extract the desired channel.}

\item{from}{Numeric. The start time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{to}{Numeric. The end time in seconds of the sample you want to analyze. Could also be relative to the end of the file (in negative values), see examples.}

\item{freq.interval}{Frequency interval to compute the RMS. Can be a vector with length two with lower and upper interval of frequencies (in Hz), or a pattern to calculate a interval (as octaves). For the last, see \link{freq.bands} function for details.}

\item{fdom.int}{Vector with length two. Vector with length two with lower and upper interval of frequencies (in Hz) to find the dominant frequency. This frequency will be used as the center the interval only if a pattern is specified in \code{freq.interval}.}

\item{wl}{Explain.}

\item{ovlp}{Explain.}

\item{CalibPosition}{Missing}

\item{CalibValue}{Canbe a value to apply of the ref value from a calib signal (specified by CalibPosition).}

\item{freq.weight}{Character. Argument passed to dBweight to indicate the weighting curve to use on the anlysis. 'A', 'B', 'C', 'D', 'ITU', and 'none' are supported. See dBweight for details. (By default: "none")}

\item{ref}{Numerical. The reference value for dB conversion. For sound in water, the common is 1 microPa, and for sound on air 20 microPa. (By default 20)}

\item{progressbar}{Logical. Activate or deactivate a progress bar with elapsed time and the last concluded file number. (By default: \code{TRUE})}
}
\description{
RMS from a sample of a sound file
}
\examples{
song.level(tham, freq.interval=c(22, 20000))

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24)

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24, freq.weight="A")

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24, freq.weight="B")

song.level(tham, freq.interval=c(22, 20000), CalibValue = 130.24, freq.weight="C")

song.level(tham, from = 3.883035, to=7.044417, freq.interval=c(22.09429, 22627.38), CalibValue = 130.24, freq.weight="A")

#Perfect fifth (music theory)
song.level(tham, fdom.int = c(800, 2000), from = 3.8, to=7, freq.interval=3/2, CalibValue = 130.24, freq.weight="A")

#Perfect fifth (music theory)
song.level(tham, fdom.int = c(800, 2000), from = 3.8, to=7, freq.interval=3/2, CalibValue = 130.24, freq.weight="A")

}
