% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soundmeter.R
\name{soundmeter}
\alias{soundmeter}
\title{Function that makes sound meter alike measurements}
\usage{
soundmeter(
  files = "wd",
  from = 0,
  to = Inf,
  CalibPosition = NULL,
  CalibValue = 0,
  ref = 20,
  fw = "none",
  bands = "octaves",
  banpass = NULL,
  tw = "fast",
  time.mess = T,
  stat.mess = T,
  channel = "left",
  saveresults = F,
  outname = NULL
)
}
\arguments{
\item{CalibPosition}{anda de mãos dadas com calib value. Pode ser negativo, positivo ou data.frame com essas combinações}

\item{CalibValue}{Anda de mãos dadas com calib position. Quando tem o position, ele é considerado o valor de referência, quando não tem o position, ele é considerado o valor de calibração.}

\item{fw}{Character. Argument passed to \code{\link[seewave]{dBweight}} to indicate the frequency weighting curve to use on the anlysis. 'A', 'B', 'C', 'D', 'ITU', and 'none' are supported. See \code{\link[seewave]{dBweight}} for details. (By default: "none")}

\item{tw}{Time weighting}

\item{time.mess}{Logical. Activate or deactivate message of time to complete the function execution. (By default: \code{TRUE})}

\item{stat.mess}{Logical. Activate or deactivate status message of the function execution. (By default: \code{TRUE})}

\item{channel}{Only "left" or "right" acepted. By default "left"}

\item{saveresults}{Logical. Set \code{TRUE} if you want to save a txt file with the results of the function execution. (By default: \code{FALSE})}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, you can specify a name to appear on the txt file name after the default name. (By default: \code{NULL})}

\item{bandpass}{falta implementar...}
}
\description{
Function that makes sound meter alike measurements
}
\details{
If your reference signal is in a separate file, we recommend you to get the \code{CalibValue} with \link{Tweighting} function. It's examples provide more details.
}
\examples{
data("tham")
soundmeter(tham, CalibValue = 130.24, tw="slow") #slow time window with calib value
soundmeter(tham, CalibValue = NULL, tw="fast") #fast time window in dBFS

}
