% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumdB.R
\name{sumdB}
\alias{sumdB}
\title{Sum with dB values}
\usage{
sumdB(x, level = "IL", na.rm = FALSE)
}
\arguments{
\item{x}{Numerical. A numeric vector or matrix with dB values.}

\item{level}{Character. Specify in what scale your data is. \code{SPL} for Sound Pressute Level or \code{IL} for Intensity Level. (By default \code{SPL})}

\item{na.rm}{Logical. Argument passed to \code{\link[base]{sum}}. Should NA be removed? (By default FALSE)}
}
\value{
A numeric value that represents the sum of x.
}
\description{
Sum with dB values
}
\details{
This function computes the sum of dB values.

This function converts your dB data to linear values (through \code{\link{dBtoLinear}} function), compute the sum, and converts the result back to dB (through \code{\link{LineartodB}} function).
}
\examples{
sumdB(c(80,60,65,62))
sumdB(c(30,30), level="IL")
sumdB(c(30,30), level="SPL")

}
\seealso{
\code{\link[seewave]{moredB}}
}
