% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timbre.R
\name{timbre}
\alias{timbre}
\title{Intensity per octave or octave thirds}
\usage{
timbre(files="wd", weighting="none", bands="thirds", saveresults=F,
       outname=NULL, Leq.calib=NULL, Calib.value=NULL, time.mess=T, stat.mess=T)
}
\arguments{
\item{files}{The audiofile to be analyzed. Can be "wd" to get all ".wav" files on the work directory, a file name (or a character containing a list of filenames) that exist in the work directory (only ".wav" files accepted), or an Wave object (or a list containing more than one Wave object). (By default: "wd")}

\item{weighting}{Character. Argument passed to \code{\link[seewave]{dBweight}} to indicate the weighting curve to use on the anlysis. 'A', 'B', 'C', 'D', 'ITU', and 'none' are supported. See \code{\link[seewave]{dBweight}} for details. (By default: "none")}

\item{bands}{Character. Choose the type of frequency band of the output. "octaves" to octaves bands intervals or "thirds" to one-third octaves bands intervals. (by deafault: "thirds")}

\item{saveresults}{Logical. Set \code{TRUE} if you want to save a txt file with the results of the function execution. (By default: \code{FALSE})}

\item{outname}{Character. If \code{saveresults} is \code{TRUE}, you can specify a name to appear on the txt file name after the default name. (By default: \code{NULL})}

\item{Leq.calib}{Numerical. The sound intensity level (in dB) that the sound in the audio file must have (by default: NULL). Can not be set if \code{Calib.value} is also set.}

\item{Calib.value}{Numerical. The calibration value returned from the analysis of a reference sound using \code{Leq.calib} (by default: NULL). Can not be set if \code{Leq.calib} is also set.}

\item{time.mess}{Logical. Activate or deactivate message of time to complete the function execution. (By default: \code{TRUE})}

\item{stat.mess}{Logical. Activate or deactivate status message of the function execution. (By default: \code{TRUE})}
}
\description{

}
\details{
Caution: You need to use an audiofile with entire values of secconds of duration to avoid bugs. Example: 35s, 60s, 19s. By default, the function will trunc your audiofile to the next entire value of seconds.

These function works only with mono audiofiles.

The audio files need to have at least 44100Hz of sampling rate.
}
\references{
Power spectrum adapted from: Carcagno, S. 2013. Basic Sound Processing with R [Blog post]. Retrieved from http://samcarcagno.altervista.org/blog/basic-sound-processing-r/

Miyara, F. 2017. Software-Based Acoustical Measurements. Springer. 429 pp. DOI: 10.1007/978-3-319-55871-4
}
\author{
Cássio Rachid Simões <cassiorachid@gmail.com>
}
